move backend code into separate dir
