+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "calyx-opt",
     "calyx-frontend",
     "calyx-stdlib",
+    "calyx-backend",
     "interp",
     "web/rust",
     "tools/data_gen",
@@ -45,6 +46,7 @@ calyx-utils = { path = "calyx-utils", version = "0.5.0" }
 calyx-ir = { path = "calyx-ir", version = "0.5.0" }
 calyx-frontend = { path = "calyx-frontend", version = "0.5.0" }
 calyx-opt = { path = "calyx-opt", version = "0.5.0" }
+calyx-backend = { path = "calyx-backend", version = "0.5.0" }
 
 [workspace.dependencies.petgraph]
 version = "0.6"
@@ -77,8 +79,8 @@ path = "src/main.rs"
 default = []
 serialize = [
     "dep:serde_with",
-    "calyx-ir/serialize",
     "dep:serde_sexpr",
+    "calyx-ir/serialize",
     "serde/rc",
 ]
 
@@ -93,18 +95,15 @@ serde.workspace = true
 serde_with = { workspace = true, optional = true }
 serde_sexpr = { workspace = true, optional = true }
 argh.workspace = true
-csv = "1.1"
-vast = "0.3.1"
 
 calyx-utils.workspace = true
 calyx-ir.workspace = true
 calyx-frontend.workspace = true
 calyx-opt.workspace = true
 
-[dependencies.quick-xml]
-version = "0.28"
-default-features = false
-features = ["serialize"]
+[dependencies.calyx-backend]
+workspace = true
+features = ["mlir", "resources", "xilinx", "sexp"]
 
 [dependencies.env_logger]
 version = "0.9.0"
