+++ calyx-backend/Cargo.toml
@@ -1,0 +1,43 @@
+[package]
+name = "calyx-backend"
+version.workspace = true
+edition.workspace = true
+description.workspace = true
+authors.workspace = true
+license-file.workspace = true
+keywords.workspace = true
+repository.workspace = true
+homepage.workspace = true
+categories.workspace = true
+readme.workspace = true
+
+[dependencies]
+log.workspace = true
+petgraph.workspace = true
+string-interner.workspace = true
+itertools.workspace = true
+linked-hash-map.workspace = true
+serde = { workspace = true }
+serde_with = { workspace = true, optional = true }
+serde_sexpr = { workspace = true, optional = true }
+smallvec.workspace = true
+
+calyx-utils.workspace = true
+calyx-frontend.workspace = true
+calyx-ir.workspace = true
+
+csv = { version = "1.1", optional = true }
+vast = "0.3.1"
+
+[dependencies.quick-xml]
+version = "0.30"
+features = ["serialize"]
+default-features = false
+optional = true
+
+[features]
+default = []
+mlir = []
+xilinx = ["dep:quick-xml"]
+resources = ["dep:csv"]
+sexp = ["dep:serde_with", "dep:serde_sexpr", "serde/rc", "calyx-ir/serialize"]
