+++ calyx-backend/src/xilinx/xml.rs
@@ -1,4 +1,4 @@
-use crate::backend::traits::Backend;
+use crate::traits::Backend;
 use calyx_ir as ir;
 use calyx_utils::CalyxResult;
 use serde::Serialize;
