--- src/backend/mod.rs
@@ -1,7 +1,0 @@
-//! Backends for the Calyx compiler.
-pub mod mlir;
-pub mod resources;
-pub mod sexp;
-pub mod traits;
-pub mod verilog;
-pub mod xilinx;
